feat: add gcs image uploads to rich text editor
