+++ packages/root-cms/package.json
@@ -75,7 +75,10 @@
     "@blinkk/root": "workspace:*",
     "@editorjs/editorjs": "^2.28.2",
     "@editorjs/header": "^2.8.1",
+    "@editorjs/image": "^2.9.0",
     "@editorjs/list": "^1.9.0",
+    "@editorjs/nested-list": "^1.4.2",
+    "@editorjs/raw": "^2.5.0",
     "@editorjs/table": "^2.3.0",
     "@emotion/react": "^11.10.5",
     "@firebase/app-compat": "^0.2.19",
