+++ packages/root-cms/ui/components/RichTextEditor/RichTextEditor.css
@@ -1,11 +1,43 @@
 .RichTextEditor {
-  border: 1px solid var(--color-border);
+  border: 1px solid #ced4da;
   padding: 4px 10px;
+}
+
+.RichTextEditor .codex-editor__redactor {
+  padding-bottom: 150px !important;
 }
 
 .RichTextEditor .cdx-list__item {
   padding-top: 0;
   padding-bottom: 0;
+}
+
+.RichTextEditor .image-tool__image {
+  margin-bottom: 0;
+}
+
+.RichTextEditor .image-tool__image-picture {
+  background: #f5f5f5;
+  border: 1px solid #dedede;
+  aspect-ratio: 16/9;
+  width: 100%;
+  padding: 10px;
+  object-fit: contain;
+  object-position: center;
+}
+
+.RichTextEditor .image-tool__image::after {
+  content: 'Image alt text';
+  display: block;
+  margin-top: 6px;
+  font-weight: 500;
+}
+
+.RichTextEditor .image-tool__caption {
+  font-family: inherit;
+  border: 1px solid #ced4da;
+  padding: 8px 10px;
+  margin-top: 4px;
 }
 
 .RichTextEditor .tc-table {
